+++ src/lib/services/txService.ts
@@ -10,6 +10,7 @@ import {
   CELATONE_QUERY_KEYS,
   useBaseApiRoute,
   useCelatoneApp,
+  useInitia,
   useMoveConfig,
   useWasmConfig,
 } from "lib/app-provider";
@@ -18,9 +19,6 @@ import {
   getTxsCountByAddress,
   getTxsCountByPoolId,
   getTxsByPoolIdPagination,
-  getBlockTransactionCountByHeightQueryDocument,
-  getBlockTransactionsByHeightQueryDocument,
-  getModuleTransactionsQueryDocument,
   getModuleTransactionsCountQueryDocument,
 } from "lib/query";
 import { createQueryFnWithTimeout } from "lib/query-utils";
@@ -32,7 +30,7 @@ import type {
   Message,
   PoolTxFilter,
   Nullable,
-  HumanAddr,
+  HexAddr,
 } from "lib/types";
 import { ActionMsgType, MsgFurtherAction } from "lib/types";
 import {
@@ -45,12 +43,20 @@ import {
 } from "lib/utils";
 
 import { usePoolTxExpression, useTxExpression } from "./expression";
-import type { AccountTxsResponse, TxResponse, TxsResponse } from "./tx";
+import type {
+  AccountTxsResponse,
+  ModuleTxsResponse,
+  TxResponse,
+  TxsResponse,
+  BlockTxsResponse,
+} from "./tx";
 import {
   getTxs,
   getTxsByAddress,
   queryTxData,
   getAPITxsCountByAddress,
+  getTxsByModule,
+  getTxsByBlockHeight,
 } from "./tx";
 
 export interface TxData extends TxResponse {
@@ -87,20 +93,23 @@ export const useTxData = (
 
 export const useTxsByAddress = (
   address: Addr,
+  search: Option<string>,
   isSigner: Option<boolean>,
   txFilters: TxFilters,
   offset: number,
   limit: number
-): UseQueryResult<AccountTxsResponse> => {
+) => {
   const endpoint = useBaseApiRoute("accounts");
   const { enabled: isWasm } = useWasmConfig({ shouldRedirect: false });
   const { enabled: isMove } = useMoveConfig({ shouldRedirect: false });
+  const isInitia = useInitia();
 
-  return useQuery(
+  return useQuery<AccountTxsResponse>(
     [
       CELATONE_QUERY_KEYS.TXS_BY_ADDRESS,
       endpoint,
       address,
+      search,
       isSigner,
       JSON.stringify(txFilters),
       limit,
@@ -112,12 +121,14 @@ export const useTxsByAddress = (
       getTxsByAddress(
         endpoint,
         address,
+        search,
         isSigner,
         txFilters,
         limit,
         offset,
         isWasm,
-        isMove
+        isMove,
+        isInitia
       ),
     { retry: 1, refetchOnWindowFocus: false }
   );
@@ -191,6 +202,8 @@ export const useTxsByAddressPagination = (
           ),
           isIbc: transaction.transaction.is_ibc,
           isInstantiate: transaction.transaction.is_instantiate ?? false,
+          // TODO: use API
+          isOpinit: false,
         }))
       );
   }, [
@@ -274,20 +287,31 @@ export const useTxsCountByAddress = ({
 // TODO: this will replace useTxsCountByAddress
 export const useAPITxsCountByAddress = (
   address: Addr,
+  search: Option<string>,
   isSigner: Option<boolean>,
   txFilters: TxFilters
 ) => {
   const endpoint = useBaseApiRoute("accounts");
+  const { enabled: wasmEnable } = useWasmConfig({ shouldRedirect: false });
 
   return useQuery(
     [
       CELATONE_QUERY_KEYS.API_TXS_COUNT_BY_ADDRESS,
       endpoint,
       address,
+      search,
       isSigner,
       JSON.stringify(txFilters),
     ],
-    async () => getAPITxsCountByAddress(endpoint, address, isSigner, txFilters),
+    async () =>
+      getAPITxsCountByAddress(
+        endpoint,
+        address,
+        search,
+        isSigner,
+        txFilters,
+        wasmEnable
+      ),
     { retry: 1, refetchOnWindowFocus: false }
   );
 };
@@ -322,6 +346,7 @@ export const useTxsByPoolIdPagination = (
           furtherAction: MsgFurtherAction.NONE,
           isIbc: transaction.transaction.is_ibc,
           isInstantiate: false,
+          isOpinit: false,
         }))
       );
   }, [expression, indexerGraphClient, offset, pageSize]);
@@ -379,163 +404,65 @@ export const useTxsCountByPoolId = (
 export const useTxs = (
   limit: number,
   offset: number,
-  options: Pick<
-    UseQueryOptions<TxsResponse, Error>,
-    "onSuccess" | "onError"
-  > = {}
-): UseQueryResult<TxsResponse> => {
+  options: Pick<UseQueryOptions<TxsResponse>, "onSuccess"> = {}
+) => {
   const endpoint = useBaseApiRoute("txs");
   const { enabled: wasmEnable } = useWasmConfig({ shouldRedirect: false });
   const { enabled: moveEnable } = useMoveConfig({ shouldRedirect: false });
+  const isInitia = useInitia();
 
-  return useQuery(
-    [CELATONE_QUERY_KEYS.TXS, endpoint, limit, offset, wasmEnable, moveEnable],
-    async () => getTxs(endpoint, limit, offset, wasmEnable, moveEnable),
+  return useQuery<TxsResponse>(
+    [
+      CELATONE_QUERY_KEYS.TXS,
+      endpoint,
+      limit,
+      offset,
+      wasmEnable,
+      moveEnable,
+      isInitia,
+    ],
+    async () =>
+      getTxs(endpoint, limit, offset, wasmEnable, moveEnable, isInitia),
     { ...options, retry: 1, refetchOnWindowFocus: false }
   );
 };
 
-export const useTxsByBlockHeightPagination = (
+export const useTxsByBlockHeight = (
   height: number,
   limit: number,
-  offset: number
-): UseQueryResult<Transaction[]> => {
-  const { indexerGraphClient } = useCelatoneApp();
+  offset: number,
+  options: Pick<UseQueryOptions<BlockTxsResponse>, "onSuccess"> = {}
+) => {
+  const endpoint = useBaseApiRoute("blocks");
   const { enabled: wasmEnable } = useWasmConfig({ shouldRedirect: false });
   const { enabled: moveEnable } = useMoveConfig({ shouldRedirect: false });
+  const isInitia = useInitia();
 
-  const queryFn = useCallback(
-    async () =>
-      indexerGraphClient
-        .request(getBlockTransactionsByHeightQueryDocument, {
-          limit,
-          offset,
-          height,
-          isWasm: wasmEnable,
-          isMove: moveEnable,
-        })
-        .then<Transaction[]>(({ transactions }) =>
-          transactions.map<Transaction>((transaction) => ({
-            hash: parseTxHash(transaction.hash),
-            messages: snakeToCamel(transaction.messages),
-            sender: transaction.account.address as Addr,
-            isSigner: false,
-            height,
-            created: parseDate(transaction.block.timestamp),
-            success: transaction.success,
-            actionMsgType: getActionMsgType([
-              transaction.is_execute,
-              transaction.is_instantiate,
-              transaction.is_send,
-              transaction.is_store_code,
-              transaction.is_migrate,
-              transaction.is_update_admin,
-              transaction.is_clear_admin,
-            ]),
-            furtherAction: MsgFurtherAction.NONE,
-            isIbc: transaction.is_ibc,
-            isInstantiate: transaction.is_instantiate ?? false,
-          }))
-        ),
-    [height, limit, offset, indexerGraphClient, wasmEnable, moveEnable]
-  );
-
-  return useQuery(
+  return useQuery<BlockTxsResponse>(
     [
-      CELATONE_QUERY_KEYS.TXS_BY_BLOCK_HEIGHT_PAGINATION,
-      indexerGraphClient,
+      CELATONE_QUERY_KEYS.TXS_BY_BLOCK_HEIGHT,
+      endpoint,
       limit,
       offset,
       height,
+      wasmEnable,
+      moveEnable,
+      isInitia,
     ],
-    queryFn,
-    {
-      keepPreviousData: true,
-      enabled: !!height,
-    }
-  );
-};
-
-export const useTxsCountByBlockHeight = (
-  height: number
-): UseQueryResult<Option<number>> => {
-  const { indexerGraphClient } = useCelatoneApp();
-
-  const queryFn = useCallback(
     async () =>
-      indexerGraphClient
-        .request(getBlockTransactionCountByHeightQueryDocument, {
-          height,
-        })
-        .then(
-          ({ transactions_aggregate }) =>
-            transactions_aggregate.aggregate?.count
-        ),
-    [height, indexerGraphClient]
-  );
-
-  return useQuery(
-    [CELATONE_QUERY_KEYS.TXS_BY_BLOCK_HEIGHT_COUNT, indexerGraphClient, height],
-    queryFn,
+      getTxsByBlockHeight(
+        endpoint,
+        height,
+        limit,
+        offset,
+        wasmEnable,
+        moveEnable,
+        isInitia
+      ),
     {
+      ...options,
       keepPreviousData: true,
       enabled: !!height,
-    }
-  );
-};
-
-export const useModuleTxsByPagination = ({
-  moduleId,
-  pageSize,
-  offset,
-}: {
-  moduleId: Option<Nullable<number>>;
-  pageSize: number;
-  offset: number;
-}) => {
-  const { indexerGraphClient } = useCelatoneApp();
-
-  const queryFn = async () => {
-    if (!moduleId) return [];
-    return indexerGraphClient
-      .request(getModuleTransactionsQueryDocument, {
-        moduleId,
-        pageSize,
-        offset,
-      })
-      .then<Transaction[]>(({ module_transactions }) =>
-        module_transactions.map<Transaction>((transaction) => ({
-          hash: parseTxHash(transaction.transaction.hash),
-          messages: snakeToCamel(transaction.transaction.messages),
-          sender: transaction.transaction.account.address as HumanAddr,
-          height: transaction.block.height,
-          created: parseDate(transaction.block.timestamp),
-          success: transaction.transaction.success,
-          actionMsgType: getActionMsgType([
-            transaction.transaction.is_send,
-            // TODO: handle more action Move msg type
-          ]),
-          furtherAction: MsgFurtherAction.NONE,
-          isSigner: false,
-          isIbc: transaction.transaction.is_ibc,
-          isInstantiate: false,
-        }))
-      );
-  };
-
-  return useQuery(
-    [
-      CELATONE_QUERY_KEYS.MODULE_TXS,
-      indexerGraphClient,
-      moduleId,
-      pageSize,
-      offset,
-    ],
-    createQueryFnWithTimeout(queryFn),
-    {
-      enabled: Boolean(moduleId),
-      retry: 1,
-      refetchOnWindowFocus: false,
     }
   );
 };
@@ -563,5 +490,42 @@ export const useModuleTxsCount = (moduleId: Option<Nullable<number>>) => {
       retry: 1,
       refetchOnWindowFocus: false,
     }
+  );
+};
+
+export const useTxsByModule = (
+  address: HexAddr,
+  moduleName: string,
+  offset: number,
+  limit: number
+) => {
+  const endpoint = useBaseApiRoute("move");
+  const { enabled: isWasm } = useWasmConfig({ shouldRedirect: false });
+  const { enabled: isMove } = useMoveConfig({ shouldRedirect: false });
+  const isInitia = useInitia();
+
+  return useQuery<ModuleTxsResponse>(
+    [
+      CELATONE_QUERY_KEYS.MODULE_TXS,
+      endpoint,
+      address,
+      moduleName,
+      limit,
+      offset,
+      isWasm,
+      isMove,
+    ],
+    async () =>
+      getTxsByModule(
+        endpoint,
+        address,
+        moduleName,
+        limit,
+        offset,
+        isWasm,
+        isMove,
+        isInitia
+      ),
+    { retry: 1, refetchOnWindowFocus: false }
   );
 };